tasks/119 r5: canonical declaration bytes + whole-frame ambiguity guards (PR #120 r4 pass)
Address the 5 findings from the codex pass on 3f12f0e:

- [P1] Canonicalize wire-v2 declaration order. encode_v2_declaration serializes
  points sorted by (namespace, local), so a caller's incidental order (e.g. from a
  HashMap) never reaches the emitted or persisted original_declaration bytes —
  host-order nondeterminism must not touch recorded bytes (rule 4).
- [P1] Detect duplicate JSON keys recursively. serde_json::Value silently keeps
  the last of a repeated key at ANY depth (two `maximize` fields would let
  true-then-false normalize as Min). A DupCheck visitor walks the whole frame;
  any duplicate key => Payload::Unknown. Robust under arbitrary_precision (a number
  is a single-key map, never a duplicate).
- [P1] Register the setup lifecycle identity in SdkSchema. A valid antithesis_setup
  now merges a fixed occurrence entry, so schema.entry(event.id) is present for
  setup events like assertions/guidance.
- [P2] Reject v2-only state ops (min/accumulate, op bytes 2/3) under a v1 or
  declaration-less stream — they stay raw rather than fabricating a state update
  from future/malformed evidence; accepted only under a BinaryV2 source.
- [P2] Prevent site-coordinate truncation. SiteId.begin_line/begin_column are now
  u64, so an untrusted coordinate above u32::MAX is preserved exactly, not wrapped
  into a colliding site.

New property tests: shuffled declaration points encode byte-identically; any
injected duplicate field => Unknown. Plus unit tests for nested dup, large numeric
token (no false dup), setup schema entry, v1 op gating, oversized site coordinate.
97 tests + proptests; build/nextest/clippy -D/fmt/deny green; public-api snapshot
refreshed (SiteId coords u64).

+++ dissonance/sdk-events/src/event.rs
@@ -53,10 +53,11 @@ pub struct SiteId {
     pub file: String,
     /// The enclosing function/class path.
     pub function: String,
-    /// The 1-based line of the assertion.
-    pub begin_line: u32,
-    /// The 1-based column of the assertion.
-    pub begin_column: u32,
+    /// The 1-based line of the assertion. `u64` so an untrusted coordinate is
+    /// preserved exactly rather than truncated into a colliding site.
+    pub begin_line: u64,
+    /// The 1-based column of the assertion (`u64` for the same reason).
+    pub begin_column: u64,
 }
 
 /// The normalized value an event carries. Occurrence and state payloads are kept
